fix(Action): do not run cache ops if build returns null
Addresses an issue with some Actions, such as the MessageCreate, where guards would return `null`.

For example let's say the bot receives a MessageCreate event from an uncached channel, we cannot handle it accordingly, as it'd break Action#cache, so we return `null`.

Doing so allows us to freely return a value that will not be handled in the Action's cache layer, making it safer and more robust.